Add WSLC volume cmdlets, tab completion

+++ src/WSLC.PowerShell/Cmdlets/RemoveContainerImage.cs
@@ -5,6 +5,7 @@ using WSLC.PowerShell.Support;
 namespace WSLC.PowerShell.Cmdlets;
 
 [Cmdlet(VerbsCommon.Remove, "ContainerImage",
+        SupportsShouldProcess = true,
         DefaultParameterSetName = CommonParameterSetNames.Default)]
 public class RemoveContainerImage : ImageOperationCmdlet
 {
@@ -14,6 +15,11 @@ public class RemoveContainerImage : ImageOperationCmdlet
     {
         foreach (var imageName in ParameterResolvers.GetImageNames(Image, ImageIdOrName))
         {
+            if (!ShouldProcess(imageName, "Remove image"))
+            {
+                continue;
+            }
+
             WslcSession.DeleteImage(imageName);
         }
